requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mongoengine>=0.29
python-decouple>=3.8
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
python-dotenv>=1.0
pytest>=8.3,<9.0
pytest-django>=4.10,<5.0
requests>=2.32.3
fastapi>=0.115.0
httpx>=0.27.2
python-multipart>=0.0.9
python-json-logger>=2.0.7
drf-spectacular>=0.27.2
django-ratelimit>=4.1.0
django-redis>=5.4.0